MEED-345 : adding a refund drawer instead of the refund modal
